Make defaults.write os-conditional in bootstrap.fish

+++ bootstrap.fish
@@ -40,6 +40,8 @@ echo "[user]
 echo "Add name and email to ~/.gitconfig"
 
 # mac only
-defaults write com.apple.dock autohide-delay -float 0 && defaults write com.apple.dock autohide-time-modifier -float 0.4 && killall Dock
-defaults write com.apple.Preview ApplePersistenceIgnoreState YES
+if test uname = "Darwin"
+  defaults write com.apple.dock autohide-delay -float 0 && defaults write com.apple.dock autohide-time-modifier -float 0.4 && killall Dock
+  defaults write com.apple.Preview ApplePersistenceIgnoreState YES
+end
